requirements.txt
cffi
cloudpickle
cycler
dask
decorator
imageio
kiwisolver
matplotlib
networkx
numpy
pandas
Pillow
pycparser
pygit
pyparsing
python-dateutil
pytz
PyWavelets
PyYAML
scikit-image
scikit-learn
scipy
six
toolz
torch
torchvision
tqdm
imageio-ffmpeg
opencv-python
face_alignment
pyfakewebcam